requirements.txt
Django==1.11.1
psycopg2==2.7.1
dj-database-url==0.4.2
gunicorn==19.7.1
whitenoise==3.3.0
djangorestframework==3.6.3
PyYAML==3.12
marshmallow==3.0.0b2
